// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/Student.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Student
 */
export interface Student {
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    name: string;
    /**
     * 
     * @type {Date}
     * @memberof Student
     */
    birthday?: Date;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    school?: string;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    currentStageId: string;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    currentGradeId: string;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    remark?: string;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    studentCode: string;
    /**
     * 
     * @type {number}
     * @memberof Student
     */
    version: number;
}

/**
 * Check if a given object implements the Student interface.
 */
export function instanceOfStudent(value: object): value is Student {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('currentStageId' in value) || value['currentStageId'] === undefined) return false;
    if (!('currentGradeId' in value) || value['currentGradeId'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('studentCode' in value) || value['studentCode'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    return true;
}

export function StudentFromJSON(json: any): Student {
    return StudentFromJSONTyped(json, false);
}

export function StudentFromJSONTyped(json: any, ignoreDiscriminator: boolean): Student {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'birthday': json['birthday'] == null ? undefined : (new Date(json['birthday'])),
        'school': json['school'] == null ? undefined : json['school'],
        'currentStageId': json['currentStageId'],
        'currentGradeId': json['currentGradeId'],
        'remark': json['remark'] == null ? undefined : json['remark'],
        'id': json['id'],
        'studentCode': json['studentCode'],
        'version': json['version'],
    };
}

export function StudentToJSON(json: any): Student {
    return StudentToJSONTyped(json, false);
}

export function StudentToJSONTyped(value?: Student | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'birthday': value['birthday'] == null ? undefined : ((value['birthday']).toISOString().substring(0,10)),
        'school': value['school'],
        'currentStageId': value['currentStageId'],
        'currentGradeId': value['currentGradeId'],
        'remark': value['remark'],
        'id': value['id'],
        'studentCode': value['studentCode'],
        'version': value['version'],
    };
}